+++ src/Piece/PieceForm/PieceForm.test.js
@@ -1,11 +1,8 @@
 import moment from 'moment';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { MemoryRouter } from 'react-router-dom';
 
 import user from '@folio/jest-config-stripes/testing-library/user-event';
 import {
   act,
-  render,
   screen,
   waitFor,
 } from '@folio/jest-config-stripes/testing-library/react';
@@ -16,18 +13,9 @@ import {
   PIECE_STATUS,
 } from '@folio/stripes-acq-components';
 
+import { renderWithRouter } from '../../../test/jest/helpers';
 import { usePieceStatusChangeLog } from '../hooks';
 import PieceForm from './PieceForm';
-
-const queryClient = new QueryClient();
-
-const wrapper = ({ children }) => (
-  <MemoryRouter>
-    <QueryClientProvider client={queryClient}>
-      {children}
-    </QueryClientProvider>
-  </MemoryRouter>
-);
 
 jest.mock('@folio/stripes-acq-components', () => {
   return {
@@ -94,12 +82,11 @@ const logs = [
 const DATE_FORMAT = 'MM/DD/YYYY';
 const today = moment();
 
-const renderPieceForm = (props = {}) => render(
+const renderPieceForm = (props = {}) => renderWithRouter(
   <PieceForm
     {...defaultProps}
     {...props}
   />,
-  { wrapper },
 );
 
 const findButton = (name) => screen.findByRole('button', { name });
