+++ test/jest/helpers/index.js
@@ -1,0 +1,1 @@
+export { default as renderWithRouter } from './renderWithRouter';
