+++ test/jest/helpers/renderWithRouter.js
@@ -1,0 +1,21 @@
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
+import { MemoryRouter } from 'react-router-dom';
+
+import { render } from '@folio/jest-config-stripes/testing-library/react';
+
+const queryClient = new QueryClient();
+
+const wrapper = ({ children }) => (
+  <MemoryRouter>
+    <QueryClientProvider client={queryClient}>
+      {children}
+    </QueryClientProvider>
+  </MemoryRouter>
+);
+
+const renderWithRouter = (component) => render(component, { wrapper });
+
+export default renderWithRouter;
